pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"


[project]
name = "omero_adi"
dynamic = ["version"]
description = "OMERO Automated Data Import (ADI) System"
authors = [{ name = "Rodrigo Rosas-Bertolini", email = "cellularimaging@amsterdamumc.nl" }]
license = { file = "LICENSE" }
requires-python = ">=3.8"
readme = "README.md"  
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = [
    "ezomero @ git+https://github.com/TheJacksonLaboratory/ezomero.git@9ea21e07f012ea17f39717925207389598d40d38",
    "pandas>=2.2.1",
    "numpy>=1.26.4",
    "openpyxl>=3.1.2",
    "python-dotenv>=1.0.1",
    "pytest>=8.2.1",
    "sqlalchemy>=2.0.28",
    "psycopg2>=2.9.7"
]

[project.optional-dependencies]
test = [
    "pytest-cov",
    "mock",
    "psycopg2-binary"
]

[tool.setuptools.packages]
find = {}  # Scan the project directory with the default parameters

[tool.setuptools_scm]

[project.urls]
"Homepage" = "https://github.com/Cellular-Imaging-Amsterdam-UMC/OMERO-Automated-Data-Import"

[tool.pytest.ini_options]
cache_dir = "/tmp/.pytest_cache"